Dockerize mixer.
This patch Dockerizes mixer. A new '--native' flag is added that, if
passed, will make mixer run on the host system the way it always did.
This defaults to false.

Before every command runs, if '--native' is false, mixer now pulls an
image file published by upstream that contains the core Clear Linux OS and
mixer toolchain for the version of Clear Linux defined in the upstreamversion
file, which mixer than uses to build a Docker image. Mixer then re-runs the
original mixer command within a container of that Docker image.

This ensures that you are always using a version of the mixer toolchain
compatible with the upstream version you are building off of. This is
required for building across format boundaries.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>