Add checksum checking for existing image file
Previously, if the Docker image file attained from upstream already
existed, the tool would skip downloading it and simply reuse it.

This patch adds checksum-checking to ensure that the existing file is
valid and unaltered. If the SHA512 checksum cannot be downloaded from
upstream, or if the checksums do not match, the existing file is ignored
and a new version is downloaded.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>