Stage mix for format bumps
This patch adds a check for non-format-bump build commands to see if the
version they are building for is outside the format range of their last
build. If so, it cancels the command and instead stages the mix for the
two format bump builds. The desired upstream version they were attempting
to build is stored in a "upstreamversion.bump" file, and the current
upstream version is replaced with the last version in the format range.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>